+++ api/src/certification/enrolment/infrastructure/repositories/certification-candidate-repository.js
@@ -3,13 +3,13 @@ import _ from 'lodash';
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { PGSQL_UNIQUE_CONSTRAINT_VIOLATION_ERROR } from '../../../../../db/pgsql-errors.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
-import { BookshelfCertificationCandidate } from '../../../../../lib/infrastructure/orm-models/CertificationCandidate.js';
-import { CertificationCandidate } from '../../../../../src/shared/domain/models/CertificationCandidate.js';
 import {
   CertificationCandidateCreationOrUpdateError,
   CertificationCandidateMultipleUserLinksWithinSessionError,
   NotFoundError,
 } from '../../../../shared/domain/errors.js';
+import { CertificationCandidate } from '../../../../shared/domain/models/CertificationCandidate.js';
+import { BookshelfCertificationCandidate } from '../../../../shared/infrastructure/orm-models/CertificationCandidate.js';
 import * as bookshelfToDomainConverter from '../../../../shared/infrastructure/utils/bookshelf-to-domain-converter.js';
 import { normalize } from '../../../../shared/infrastructure/utils/string-utils.js';
 import { CompanionPingInfo } from '../../domain/models/CompanionPingInfo.js';
