+++ api/src/shared/infrastructure/orm-models/CertificationCandidate.js
@@ -1,7 +1,7 @@
 import './Session.js';
 import './User.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'CertificationCandidate';
