+++ api/src/shared/infrastructure/orm-models/CertificationCenter.js
@@ -1,6 +1,6 @@
 import './ComplementaryCertification.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'CertificationCenter';
