+++ api/src/shared/infrastructure/orm-models/CertificationCenterMembership.js
@@ -1,7 +1,7 @@
 import './CertificationCenter.js';
 import './User.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'CertificationCenterMembership';
