+++ api/src/shared/infrastructure/orm-models/CertificationIssueReport.js
@@ -1,4 +1,4 @@
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'CertificationIssueReport';
