+++ api/src/shared/infrastructure/orm-models/Membership.js
@@ -1,14 +1,14 @@
-import './User.js';
 import './Organization.js';
+import './User.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
-const modelName = 'OrganizationLearner';
+const modelName = 'Membership';
 
-const BookshelfOrganizationLearner = Bookshelf.model(
+const BookshelfMembership = Bookshelf.model(
   modelName,
   {
-    tableName: 'organization-learners',
+    tableName: 'memberships',
     hasTimestamps: ['createdAt', 'updatedAt'],
 
     user() {
@@ -24,4 +24,4 @@ const BookshelfOrganizationLearner = Bookshelf.model(
   },
 );
 
-export { BookshelfOrganizationLearner };
+export { BookshelfMembership };
