+++ api/src/shared/infrastructure/orm-models/Organization.js
@@ -1,6 +1,6 @@
 import './Tag.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'Organization';
