+++ api/src/shared/infrastructure/orm-models/Tag.js
@@ -1,4 +1,4 @@
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'Tag';
