+++ api/src/shared/infrastructure/orm-models/UserOrgaSettings.js
@@ -1,7 +1,7 @@
 import './User.js';
 import './Organization.js';
 
-import { Bookshelf } from '../bookshelf.js';
+import { Bookshelf } from '../../../../lib/infrastructure/bookshelf.js';
 
 const modelName = 'UserOrgaSettings';
