+++ api/tests/acceptance/scripts/add-many-divisions-and-students-to-sco-organization_test.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../lib/infrastructure/bookshelf.js';
-import { BookshelfOrganizationLearner } from '../../../lib/infrastructure/orm-models/OrganizationLearner.js';
 import { addManyDivisionsAndStudentsToScoCertificationCenter } from '../../../scripts/data-generation/add-many-divisions-and-students-to-sco-organization.js';
+import { BookshelfOrganizationLearner } from '../../../src/shared/infrastructure/orm-models/OrganizationLearner.js';
 import { databaseBuilder, expect } from '../../test-helper.js';
 
 describe('Acceptance | Scripts | add-many-divisions-and-students-to-sco-organization', function () {
