+++ api/tests/acceptance/scripts/add-many-students-to-sco-certification-center_test.js
@@ -1,5 +1,5 @@
-import { BookshelfOrganizationLearner } from '../../../lib/infrastructure/orm-models/OrganizationLearner.js';
 import { addManyStudentsToScoCertificationCenter } from '../../../scripts/data-generation/add-many-students-to-sco-certification-center.js';
+import { BookshelfOrganizationLearner } from '../../../src/shared/infrastructure/orm-models/OrganizationLearner.js';
 import { databaseBuilder, expect } from '../../test-helper.js';
 
 describe('Acceptance | Scripts | add-many-students-to-sco-certification-centers.js', function () {
