+++ api/tests/integration/scripts/create-certification-center-memberships-from-organization-admins_test.js
@@ -1,4 +1,3 @@
-import { BookshelfCertificationCenterMembership } from '../../../lib/infrastructure/orm-models/CertificationCenterMembership.js';
 import {
   createCertificationCenterMemberships,
   fetchCertificationCenterMembershipsByExternalId,
@@ -7,6 +6,7 @@ import {
   prepareDataForInsert,
 } from '../../../scripts/create-certification-center-memberships-from-organization-admins.js';
 import { Membership } from '../../../src/shared/domain/models/Membership.js';
+import { BookshelfCertificationCenterMembership } from '../../../src/shared/infrastructure/orm-models/CertificationCenterMembership.js';
 import { databaseBuilder, expect } from '../../test-helper.js';
 
 describe('Integration | Scripts | create-certification-center-memberships-from-organization-admins.js', function () {
