+++ api/tests/shared/integration/infrastructure/utils/bookshelf-to-domain-converter_test.js
@@ -1,13 +1,13 @@
-import { BookshelfCampaign } from '../../../../../lib/infrastructure/orm-models/Campaign.js';
-import { BookshelfCampaignParticipation } from '../../../../../lib/infrastructure/orm-models/CampaignParticipation.js';
-import { BookshelfOrganization } from '../../../../../lib/infrastructure/orm-models/Organization.js';
-import { BookshelfUser } from '../../../../../lib/infrastructure/orm-models/User.js';
 import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
 import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
 import { Tag } from '../../../../../src/organizational-entities/domain/models/Tag.js';
 import { KnowledgeElement } from '../../../../../src/shared/domain/models/KnowledgeElement.js';
 import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
 import { TargetProfile } from '../../../../../src/shared/domain/models/TargetProfile.js';
+import { BookshelfCampaign } from '../../../../../src/shared/infrastructure/orm-models/Campaign.js';
+import { BookshelfCampaignParticipation } from '../../../../../src/shared/infrastructure/orm-models/CampaignParticipation.js';
+import { BookshelfOrganization } from '../../../../../src/shared/infrastructure/orm-models/Organization.js';
+import { BookshelfUser } from '../../../../../src/shared/infrastructure/orm-models/User.js';
 import * as bookshelfToDomainConverter from '../../../../../src/shared/infrastructure/utils/bookshelf-to-domain-converter.js';
 import { databaseBuilder, expect } from '../../../../test-helper.js';
